# gen/openapi/noebs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "noebs/wallet/v1/wallet.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/wallet/currencies": {
      "get": {
        "operationId": "WalletPublicService_ListCurrenciesPublic",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListCurrenciesPublicResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "WalletPublicService"
        ]
      }
    },
    "/wallet/currencies/{currencyCode}": {
      "get": {
        "operationId": "WalletPublicService_GetCurrencyPublic",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetCurrencyPublicResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "currencyCode",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "WalletPublicService"
        ]
      }
    },
    "/wallet/deposits": {
      "post": {
        "operationId": "WalletPublicService_RequestDeposit",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RequestDepositResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RequestDepositRequest"
            }
          }
        ],
        "tags": [
          "WalletPublicService"
        ]
      }
    },
...
    "/wallet/fx/quotes": {
      "post": {
        "operationId": "WalletPublicService_QuoteConversionPublic",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1QuoteConversionPublicResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1QuoteConversionPublicRequest"
            }
          }
        ],
        "tags": [
          "WalletPublicService"
        ]
      }
    },
    "/wallet/fx/quotes/{quoteId}": {
      "get": {
        "operationId": "WalletPublicService_GetConversionQuotePublic",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetConversionQuotePublicResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "quoteId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "WalletPublicService"
        ]
      }
    },
    "/wallet/fx/sources": {
      "get": {
        "operationId": "WalletPublicService_ListFXSourcesPublic",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListFXSourcesPublicResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "WalletPublicService"
        ]
      }
    },
    "/wallet/methods": {
      "get": {
        "operationId": "WalletPublicService_ListPaymentMethodsPublic",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListPaymentMethodsPublicResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "tenantId",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "region",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "amount",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "WalletPublicService"
        ]
      }
    },
    "/wallet/money/format": {
      "post": {
        "operationId": "WalletPublicService_FormatMoneyPublic",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1FormatMoneyPublicResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1FormatMoneyPublicRequest"
            }
          }
        ],
        "tags": [
          "WalletPublicService"
        ]
      }
    },
    "/wallet/money/parse": {
      "post": {
        "operationId": "WalletPublicService_ParseMoneyPublic",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ParseMoneyPublicResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ParseMoneyPublicRequest"
            }
          }
        ],
        "tags": [
          "WalletPublicService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1FormatMoneyPublicRequest": {
      "type": "object",
      "properties": {
        "tenantId": {
          "type": "string"
        },
        "currencyCode": {
          "type": "string"
        },
        "minorUnits": {
          "type": "string"
        },
        "roundingMode": {
          "type": "string"
        },
        "currencyUnitVersionId": {
          "type": "string",
          "format": "int64",
          "description": "Exact immutable unit definition that originally interpreted minor_units.\nHistorical and inactive definitions remain valid for rendering."
        }
      }
    },
    "v1FormatMoneyPublicResponse": {
      "type": "object",
      "properties": {
        "money": {
          "$ref": "#/definitions/v1MoneyAmount"
        }
      }
    },
...
    "v1GetConversionQuotePublicResponse": {
      "type": "object",
      "properties": {
        "quote": {
          "$ref": "#/definitions/v1ConversionQuote"
        }
      }
    },
    "v1GetCurrencyPublicResponse": {
      "type": "object",
      "properties": {
        "currency": {
          "$ref": "#/definitions/v1CurrencyUnit"
        }
      }
    },
...
    "v1ListCurrenciesPublicResponse": {
      "type": "object",
      "properties": {
        "currencies": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1CurrencyUnit"
          }
        }
      }
    },
    "v1ListFXSourcesPublicResponse": {
      "type": "object",
      "properties": {
        "sources": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1FXSource"
          }
        }
      }
    },
...
    "v1ListPaymentMethodsPublicResponse": {
      "type": "object",
      "properties": {
        "methods": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1PaymentMethod"
          }
        }
      }
    },
...
    "v1ParseMoneyPublicRequest": {
      "type": "object",
      "properties": {
        "tenantId": {
          "type": "string"
        },
        "currencyCode": {
          "type": "string"
        },
        "majorUnits": {
          "type": "string"
        },
        "roundingMode": {
          "type": "string"
        },
        "currencyUnitVersionId": {
          "type": "string",
          "format": "int64",
          "description": "Exact immutable unit definition selected by the client from the currency\ncatalog. Parsing rejects inactive or not-yet-effective definitions."
        }
      }
    },
    "v1ParseMoneyPublicResponse": {
      "type": "object",
      "properties": {
        "money": {
          "$ref": "#/definitions/v1MoneyAmount"
        }
      }
    },
...
    "v1QuoteConversionPublicRequest": {
      "type": "object",
      "properties": {
        "tenantId": {
          "type": "string"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "sourceCode": {
          "type": "string"
        },
        "baseCurrency": {
          "type": "string"
        },
        "quoteCurrency": {
          "type": "string"
        },
        "inputMinorUnits": {
          "type": "string"
        },
        "side": {
          "type": "string"
        },
        "baseCurrencyUnitVersionId": {
          "type": "string",
          "format": "int64",
          "description": "Exact immutable units used for the input and output amounts. A quote never\nresolves a \"current\" unit from currency code alone."
        },
        "quoteCurrencyUnitVersionId": {
          "type": "string",
          "format": "int64"
        },
        "idempotencyKey": {
          "type": "string",
          "description": "Required opaque retry key. Reusing a key with the same semantic request\nreturns the original quote; reusing it for a different request fails."
        }
      }
    },
    "v1QuoteConversionPublicResponse": {
      "type": "object",
      "properties": {
        "quote": {
          "$ref": "#/definitions/v1ConversionQuote"
        }
      }
    },
...
    "v1RequestDepositRequest": {
      "type": "object",
      "properties": {
        "tenantId": {
          "type": "string"
        },
        "providerCode": {
          "type": "string"
        },
        "walletId": {
          "type": "string"
        },
        "amount": {
          "type": "string",
          "format": "int64"
        },
        "currency": {
          "type": "string"
        },
        "idempotencyKey": {
          "type": "string"
        },
        "metadata": {
          "type": "object"
        },
        "region": {
          "type": "string"
        }
      }
    },
    "v1RequestDepositResponse": {
      "type": "object",
      "properties": {
        "workflowId": {
          "type": "string"
        },
        "runId": {
          "type": "string"
        }
      }
    },
...
  }
}